Create V3 Ticketing Program
Ticketing Program that checks if you can ride the rollercoaster if you are 120cm or taller. Then it takes into account your age for the amount to charge and with Version 3 now it checks for people in between ages 45 to 55 because they can get a free ticket. Lastly it checks if you want a photo which adds up to your final bill.